Convert the permissions tests over to mocha & chai.
Fixes #3188

Nothing of substance has really changed in the bodies of the tests
themselves; they are the same except for some find & replace changes
like:

    s/test.isTrue/chai.assert.isOk/g

...but they pass now.

This also adds explicit imports & exports for the `SandstormPermissions`
global, and adds a side-effect only import of another module to the
tests. This was done now because:

* Unless we pass `--full-app` to `meteor test`, meteor does not
  implicitly load all of our modules.
* Passing `--full-app` runs into problems trying to start the smtp
  server on startup, because our code is written to assume it was passed
  FDs by the gateway, but `meteor test` is running it in a more
  conventional fashion.
* It was easier to just fix the imports (which is what we want to do
  long term anyway) than to fight with `--full-app`.

This also turns the tests back on in CI.
