+++ shell/.meteor/packages
@@ -43,3 +43,4 @@ google-oauth@1.3.0
 dynamic-import@0.5.1
 tap:i18n
 typescript@3.7.5
+meteortesting:mocha
