+++ shell/.meteor/versions
@@ -64,6 +64,9 @@ mdg:validation-error@0.5.1
 meteor@1.9.3
 meteor-base@1.4.0
 meteorspark:util@0.2.0
+meteortesting:browser-tests@1.3.3
+meteortesting:mocha@2.0.0
+meteortesting:mocha-core@7.0.1
 minifier-css@1.5.0
 minifier-js@2.6.0
 minimongo@1.5.0
