+++ shell/imports/sandstorm-ui-powerbox/powerbox-server.js
@@ -21,6 +21,7 @@ import { Meteor } from "meteor/meteor";
 import { check } from "meteor/check";
 import { _ } from "meteor/underscore";
 
+import { SandstormPermissions } from "/imports/sandstorm-permissions/permissions.js";
 import Capnp from "/imports/server/capnp.js";
 
 import { waitPromise } from '../server/async-helpers.js';
