+++ shell/imports/server/gateway-router.js
@@ -26,6 +26,7 @@ import { Random } from "meteor/random";
 import Capnp from "/imports/server/capnp.js";
 import { SandstormDb } from "/imports/sandstorm-db/db.js";
 import { globalDb } from "/imports/db-deprecated.js";
+import { SandstormPermissions } from "/imports/sandstorm-permissions/permissions.js";
 
 import { makeHackSessionContext } from "/imports/server/hack-session.js";
