+++ shell/imports/server/grain-server.js
@@ -25,6 +25,7 @@ import { send as sendEmail } from "/imports/server/email.js";
 import { inMeteor, waitPromise } from "/imports/server/async-helpers.js";
 import { SandstormDb } from "/imports/sandstorm-db/db.js";
 import { globalDb } from "/imports/db-deprecated.js";
+import { SandstormPermissions } from "/imports/sandstorm-permissions/permissions.js";
 
 const ROOT_URL = process.env.ROOT_URL;
