+++ shell/package.json
@@ -37,6 +37,7 @@
   },
   "devDependencies": {
     "babel-eslint": "^10.1.0",
+    "chai": "^4.2.0",
     "eslint": "^6.8.0"
   }
 }
